overhaul tileserver, almost working
